Fix run-windows NuGet Errors in CI By Using Deterministic Binary (#5330)
* Try to Fix CI NuGet Errors

This change makes us use a deterministic NuGet binary that we always download. We remove the NuGetPath option which isn't exposed as a CLI option, and don't trust any local NuGet exe because we can't guarantee version. Instead of wrapping the entire logic in a recursive retry when downloading the binary, we isolate retries to just downloading the binary itself. We remove some of the blanket catches, command output parsing, and delegation to PowerShell instead of JS functions that were somehow leading to executing the whole command when trying to locate it?

Haven't tested locally. Will tell what CI thinks.

* Change files

* Rename changelog entry

* Close the write stream

* Https

* Make sure both streams finish sucessfully

* Update netcore version as azure devops images drop support

* Remove accidentally generated headers

Co-authored-by: Danny van Velzen 🁴 <dannyvv@microsoft.com>