Share eslint and Typescript configs across packages
We create rnw-scripts packages which export a typescript config and an eslint config. We can change individual packages to import these. We pull in a newer version of esling (and add is as a devDep to all packages), letting us include eslintignore rules in the base eslintrc.

This also leads to enforcing strict mode for all Typescript projects, enabling eslint for react-native-win32, react-native-windows-init (fixing some errors and suppressing other parts that aren't clean). ES6 module behavior also changed slightly for some projects.