Migrate the local-cli to TypeScript
Note that this depends on #5370, but I ommited its changes in this PR. This will be rebased upon master once that is merged in, but CI will fail in the meantime.

A few steps here:
- Update build logic to run tsc in multiple places in vnext
  - Also fixup clean logic whcih keeps on getting out of date
  - Bumped Node requirements for the repo to Node 12 for development so we can use flat/flatMap without lodash
- Separate local-cli code into a new directory structure
  - src for JS inputs
  - lib-commonjs for TS outputs
  - powerhsell and templates for others
  - Add a local gitignore to not wipe out template pfx files
  - Update CLI code for new paths
- Mechanical conversion from JS to TS
  - Very lax TS rules
  - Types substituted where needed or where documented
  - CommonJS moved to ES6 apart from where we export to raw JS
  - Mostly mechanical
- Update react-native-windows-init for the new changes
  - Try new JS path but fallback to the old if unavailable
  - Do a type only import from the new path so we can get type safety between packages
  - add RNW as a devDependency so that local-cli typings are built first by lerna

We are using dangerously relaxed TS rules here as a start. I'm hoping that we can incrementally add types and get to the point where we don't need to reduce strictness of local-cli.

Haven't done much of any validation here apart from seeing that we build. Will try to come up with something here.