Enforce null/undefined safety for the CLI
This change continues the TS work, and brings the CLI close to the level of strictness of other TS projects in the repo. The main change here is around the strictNullChecks rule, where we will now see compile time failures if null can be propagated when not declared.

A couple strategies were taken to fix these:
- Fix up types or init ordering to make it easy to reason about nullability
- Throw exceptions where we used to previously propagate unchecked null values
- Mark some things as allowed to be dull
- Add a smattering of the "!" operator to assert to the compiler that we know something is not null

There were a couple of other quick fixes around not publishing unneeded sources, not coercing object to strings that we log, and ensuring the RunWindows BuildArch is exact.