Run Release Note Generation Script in 0.62 Branch
We added a script to generate automatic release notes, and have it being published from our master branch. New stable branches will automatically run the script, but we need to bring the change back to 0.62 to get automated release notes for it.

Note that the script itself is serviced, so we can change format, etc in master and have it affect new releases of old branches once the CI change is present.