Replace Custom Generator Prompt With Inquirer
We have a chunk of code to syncrhonously read "y" or "n" when prompting users to ask if we should replace their existing files. @react-native-community/cli already includes inquirer for prompting duties. Add inquirer to our deps as well and use it instead of rolling our own logic.

Note that this means moving a lot of this code to be async. This is mostly mechanical, but changes the generator boundary, meaning we need to add more fallback logic to react-native-windows init.

Haven't validated this yet, but will test against master and 0.62 before merging.