Include "nuget.exe" as Package Dependency of react-native-windows (#5418)
* Include "nuget.exe" as Package Dependency of react-native-windows

Fixes #5404

The nuget.exe CLI tool is not provided as part of a Visual Studio installation, and we cannot use the MSBuild builtin functionality without PackageReference. We used to have logic in run-windows to download a NuGet binary if one isn't present, or try to find one if not intalled.

We can leverage npm to ensure we have a working version of NuGet instead. This change adds some build logic to publish NuGet to the NPM registry and ensure react-native-windows downloads the NuGet package during installation.

* Change files

* Fix typo