Create a Programmatic API for react-native-platform-override (#5429)
* Create a Programmatic API for react-native-platform-override

FIxes #4856
FIxes #4854
Fixes #5208
Fixes #5206

The crux of this change is to extract logic for the `react-native-platform-override` CLI into a programmatic API that can be used from other JS scripts (i.e. for automating integration). This layer is now testable, and end to end tests are added. When creating this layer, we fix up some of the other issues in the layer, such as manifest discovery (we now enforce they are package local), and changing commands a bit. With this now solidified we add a README describing behavior.

A few interesting tidbits:
- I discovered `yarn watch` was broken and fixed its preset (and added its scripts to other nnw-scripts consumers)
- We exclude compiled test code from our package
- `react-native-platform-override` VS code config is updated to allow debugging E2E tests and fixing it as a root workspace
-  We reduce scratch repo usage in end to end tests when they're updated
- We have big diffs in the override manifests due to merging them. The vnext one will become less hairy once we can defined ReacCopies by directory instead of by file.

* Change files