Don't try to run Jest Tests From the Repo Root
Current behavior of running Jest from the Repo root is to enumerate quite a few tests in different packages, which then fail, as some aren't Jest tests and others need configuration. We test packages independently using Lerna already, so the main case where this comes up is if someone uses the Jest VS Code plugin with the repo open. I spend some time trying to coerce the Jest multi-project options to work correctly, but couldn't get reasonable behavior out of it. Instead we now print a warning and don't try to run anything.