Improve Scratch Repo Access Patterns When Upgrading Overrides
react-native-platform-override currently upgrades overrides sequentially, for each one checking out an old version of RN from the scratch Git repo along with a new one. This is costly because checkouts are expensive, and we do multiple per upgrade.

This change makes it so that upgrades are triggered concurrently, and adds a mechanism to GitReactFileRespository to batch together requests of the same React Native version. This dramatically reduces checkout frequency.

An informal benchmarkl was done integrating 2.5 months of updates:
- Previous time: **1m20s**
- New time: **33s**