Remove ReactWindows-Universal.sln and update publish.yml to use Microsoft.React-native.sln
The publish.yml has been failing to create nuget packages because it couldn't find
the files for Microsoft.Reactnative.Manage.CodeGen to create the nuget package.
The reason was that publish.yml was building ReactWindows-Universal.sln
The original reason was to build ReactUWP which is no longer build.
After diffing the 2 solutions, the only project still in ReactWindows-Universal
not in Microsoft.ReactNative.sln is React.Windows.IntegrationTests.vcxproj but
that project is part for ReactWindows-Desktop.sln so it is covered and doesn't
need to be added to Microsoft.ReactNative.sln