Move 0.63-stable to public react-native 0.63
We choose 0.63.0 as a build that's close to our current master branch, but reduces headaches by being a non-prerelease build.

Not that there were some TextInput.js changes made upstream to move non-Android platforms to use ViewManager commands instead of setting native props. This was seemingly Fabric work that was collateral to take an iOS native component fix that doesn't apply to us. Instead of implementing ViewManager commands in the 0.63 branch that we haven't done yet in master, we just keep the old JS which sets native props for 0.63.

Validated that after integration we pass override validation, are flow clean, eslint clean, can launch RNTester using Rex and XAML.