Fix Crash Using accessibilityValue Without Web Debugging
Existing code will crash with a folly TypeError since we enforce the folly::dynamic value is an integer. Using Chakra, in AccessibilityExample we see the dynamic for min represented as a double (0.00000) and crash.

We're converting to a double anyway, so we instead rely on builtin folly type converstion instead of trying to coerce the value to an int to convert into a double again.

Note that we still see issues with the AccessibilityExample page.