Initial Version of Automatic React Native Integration Script (#5660)
* Initial Version of Automatic React Native Integration Script

This is a rough start to scripting the automatic integration of React Native. Running `yarn integrate-rn <version>` will update package.json's of packages to the new RN version, synchronize dependencies, do a three-way-merge of overrides, fix lint errors, trigger a JS build, and do a flow-check.

This is in a rough state, but is already more useful than doing things manually.

Steps for the future:
- Improve output in error cases
- Allow automating PR creation/markdown output
- Add test coverage
- Diff files to detect ViewManager API changes

* Change files

* Add missing deps

* Polish

* Fix promoteRelease

* Fix typos