Integrate 4/29 Nightly Build
Integrating a week of changes. This one was nice and gentle in terms of conflicting files.

Validated that RNTester launches. Looked through the week of commits for anything standing out as needing more attention and didn't see anything.

We also get dark mode new app screen with these changes :).

Fixes #4678