Fix Errant "Nested elements in TextInput are currently unsupported on Windows" Yellowbox Warnings (#5703)
* Fix Errant "Nested elements in TextInput are currently unsupported on Windows" Yellowbox Warnings

When standing up the TextInput RNTester page I ran into ViewManagerBase assertions related to trying to embed Text elements inside of TextInput. To avoid debug assertions which may break tests, I implemented child manipulation functions in TextInputShadowNode to Yellowbox instead. One of the functions implemented was removeAllChildren, which seems to be called in cases where there were no children to begin with. This leads to seeing "Nested elements in TextInput are currently unsupported on Windows" in cases where we shouldn't.

ViewManagerBase doesn't assert when the default removeAllChildren is called, and we shouldn't either.

* Change files