Integrate 4/29 Nightly Build  (#5702)
* Integrate 4/29 Nightly Build

Integrating a week of changes. This one was nice and gentle in terms of conflicting files.

Validated that RNTester launches. Looked through the week of commits for anything standing out as needing more attention and didn't see anything.

We also get dark mode new app screen with these changes :).

Fixes #4678

* Change files

* Integrate new version of codegen that has changes from react-native

* Change files

* Merge master branch

* new lockfile

Co-authored-by: Andrew Coates <acoates-ms@noreply.github.com>