Remove acceptsKeyboardFocus Property
This change, targetting RNW 0.64 removes the acceptsKeyboardFocus property that was deprecated in RNW 0.63 in favor of focusable. We update typings to reflect this and remove support from built-in view managers. focusable was initially added to ViewWindows and ViewWin32 typings when we had outdated typings that didn't have changes made in RN 0.62. With newer typings we can remove these (we had them before but bump to 0.63 with this change). We do keep acceptsKeyboardFocus in react-native-win32 view configs to allow n-1 version support when migrating RN versions in devmain.

Still validating. Have a separate internal change to implement focusable in NetUI.