Add verification to ensure nuget and npm version match.
This will add a targets file: Microsoft.ReactNative.VersionCheck.targets
to all nuget packages. When evaluated in msbuild it will perform a check
to see if the specified version of the react-native-windows package matches
the current nuget package.

It is challenging with a mix of `PackageReference` and `packages.json` (UWP doesn't support the former yet)
to consistenly figure out the version number from within the nuget package itself.
It would involve walking the parent folder looking for nuspec or npkg files and reverse engineering
the version number from the nupkg filename. In the nuspec case (PackageReference) it is simple as you just need to
parse the nusepc file. In the UWP case, it is complicated since there is no nuspec file by default, on a nupkg file.
The issue with that is that the filename is <id>.<version> and the dot separator is valid in both the id and the version
so there is no clear way to know which is which. The only approach would have
been to extract the zip file (with nuget semantics, it's not a straight unzip as filenames are url-encoded in the zip) and
then parse the nuspect inside...

To simplify this change takes a short-cut of baking the nuget version into the targets file during the build.