Fix bugs from acceptsKeyboardFocus to focusable deprecation  (#5818) (#5837)
* Fix bugs from acceptsKeyboardFocus to focusable deprecation

There are some other bits polished, like removing an erroneous acceptsKeyboardFocus in a ViewConfig relied on by RCTText, and removing react-native-win32 typings from ViewWin32 since its deprecation cycle looks a bit different, and it helps us fully remove the property in 0.64.

A runtime warning is added on property usage to discourage use, to help catch cases where folks aren't relying on typings for compile-time errors. This is done by intercepting properties to RCTView, which will also catch property usage on Touchables, TouchableWin32, and other composed views.

To avoid bleadthrough to RN 0.65 in Office, we remove typings from ViewWin32 now to encourage transition.

Validated we see the warning when a prop is present, and that touchables are focusable by default, Touchables will use acceptsKeyboardFocus instead of focusable when present. Plan is to backport this to 0.63 and remove props from master shortly after.

* Change files

* yarn api

* Remove another acceptsKeyboardFocus

* Force E2ETest to show colors

* Avoid duplicate NuGet Restore

* Remove forking disabling focusable, but keep forking adding testID