Remove acceptsKeyboardFocus (for real this time)
We've backported changes to 0.63 to fix bugs with focusable, and to yellowbox if acceptsKeyboardFocus is used on View or touchables. This formalizes the removal, turning the yellowbox into a redbox, removing VM support, and removing typings. Note that typings for react-native-win32 were already removed in 0.63, so there isn't the work to do there.

Because we removed forking from Touchables we will no longer see warnings if acceptsKeyboardFocus is still used on them, but that scenario should be rarer and we will have yellowboxed for it already in 0.63.

Tested we redbox if we try to use the prop