Fix Inspector in react-native-win32 0.62  (#5845)
* Fix Inspector in react-native-win32 0.62

We used to not have automatic dependency syncing. This led to an out-of-date depedency on react-devtools-core, which changed how certain globals touched by RN are represented. In effect, launching Inspector from the debug menu (and likely other places) was broken.

Manually sync dependencies with RN core. This is still in a hazy state, since we don't declare everything we should, but the issue is resolved in later versions of RNW.

* Change files

* Disable test disabled in master which tests logging capabilities of stub instance

* Change files