Prevent Clicks/Taps From TextInput to Touchables in ScrollView From Being Eaten
See #5867

This backports the upstream change made in https://github.com/facebook/react-native/pull/29798 to older versions of react-native-windows. Not going to bother to add the patching to the master branch, since we will pull in the fix automatically before 0.64.

Validates the issue no longer repros after the change in the multicolumn flatlist example.