Keep Templates in react-native-windows Package (#5893)
* Keep Templates in react-native-windows Package

Fixes #5884

Testing new app creation with `react-native-windows-init` can be done using `--useDevMode` and `yarn link`, or can be done by pointing `react-native-windows-init` to a fille-path. The former works by symlinking to the monorepo, and the latter copies the single package.

The latter workflow will not pick up changes to packages outside of react-native-windows. This includes template files that were moved outside the package.

This change moves template files back into react-native-windows to ease that workflow, but keeps the CLI as a separate package to avoid previous issues of intermingled dependencies, duplicated build logic, etc. We keep the templates in a top-level `template` directory, similar to what is done by react-native core. https://github.com/facebook/react-native/tree/master/template

This change complicates logic a bit to resolve tempate location, since we can't create a cyclical peer dependency on react-native-windows. We instead need to resolve the location of react-native-windows dynamically. We do that by searching from the project directory.

WIll validate by init tests in CI. Not sure if there is a current way to validate autolinking.

* Change files

* Remove unused import

* We can't resolve package directories directly

* Update docs