[0.63] Fix 5870: In C# CodeGen, don't check for inheritance, but do explicit type checks (#5873)
* Fix 5870: Dont check for inheritance, but do explicit type check

The fix for #5617 introduced a regression. The fix for that issue updated the code
to allow for all derived types of `IReactPromise<T>`. In this exapmle the C# compiler
found an assignability conversion between `IReadOnlyList<string>` and `IReactPromise<T>`.
Which resulted in invalid generated code that doesn't compile.
This fix was in hindsight also not the correct fix since the codegen (jsut like the
reflection based code) passes ReactPromise<T> instance. So the only two valid types
should be `IReactPromise<T>` or `ReactPromise<T>`. Which is what this fix puts in.

Fixes #5870

* Change files

Co-authored-by: Jon Thysell (JAUNTY) <jthysell@microsoft.com>