Enable @typescript-eslint/no-floating-promises and @typescript-eslint/no-misused-promises  (#5924)
* Enable @typescript-eslint/no-floating-promises

This rule will create errors if we have an unhandled promise, such as forgetting the "await" operator. Enable the rule and fix up current instances by adding "await" or using the "void" operator, as the rule suggests to signal intentional floating promises.

This rule requires type information, which requires pointing eslint config to a local project. This doesnt't work over an entire monorepo, such as when using VS Code outside of an individual package, so we split up our eslint config into a per-package one and one that can run in the package root without project-local info. The former is run by "yarn lint" during CI and the latter is shown by VS Code.

This also found some accidentally checked in build targets that we can remove.

* Change files

* Keep wrapped in conditional

* Add @typescript-eslint/no-misused-promises