Start Sharing Test Configuration Between Node Projects  (#5946)
* Start Sharing Test Configuration Between Node Projects

This change moves some of the work I did for setting up unit tests and end to end tests for react-native-platform-override to a shared location. This prepares us for enabling these sorts of tests across projects in the repo without duplicating configuration. Some work still needs to be done per-project to enable this, and some of this still needs to be documented.

Validated:
- One-click debug tests from VS Code
- Unit tests are automatically run when you type, end to end tests aren't
- yarn test will automatically detect unit tests and end to ends tests

* Change files

* Remove extra babel config