Test Runner for Integration Test Harness (#5964)
* WIP Integration Test Runner for MSRN

* Revert

* Polish

* Start Sharing Test Configuration Between Node Projects

This change moves some of the work I did for setting up unit tests and end to end tests for react-native-platform-override to a shared location. This prepares us for enabling these sorts of tests across projects in the repo without duplicating configuration. Some work still needs to be done per-project to enable this, and some of this still needs to be documented.

Validated:
- One-click debug tests from VS Code
- Unit tests are automatically run when you type, end to end tests aren't
- yarn test will automatically detect unit tests and end to ends tests

* Change files

* Fix Closing a UWP Websocket

The arguments for WebsocketModule were wrong on UWP. Fix that, and start firing closed events like we should. This is enough to get websocker integration tests passing.

This should all eventually be replaced by the new WebSocketResource code once we don't need to touch it as much.

* Make improvements to native code

* JS Code

* Separate out TestWebSocketServer

* Less Jank Synchronization

* Fixes

* Fixes

* Increase WebSocketJSExecutor Retry Count

We will often see "Prepare JS runtime timed out, Executor instance is not connected to a WebSocket endpoint." in cases where we start an app directly after starting Metro, and the bundler is still taking time to initialize.

Increase the retry count.

* Native fixes

* More JS

* Remove old folder

* Reconcile and Bump Some Dependencies

- Request consistent versions of Chalk, Ora, Semver, Yargs
- Bring in Jest 26, which seems to work correctly for our tests (early Jest 25 didn't)

* Better JS

* Fix x86 Crash on Exception

* yarn

* Higher level, custom, automatically registered tests

* Fixes

* yarn

* Change files

* Jest config

* Last minute tweaks