Run Non-UI-Thread Core Modules on the JS Thread
We invoke modules provided by the NativeModuleProvider on the JS thread as of RN 0.63. We still invoke core modules on either the UI thread (ReactUwp style) or an arbitrary background thread via MakeSerialQueueThread(). Invoke non-UI thread modules on the JS thread for consistency.

There's also a bit of renaming here, since the previous differentiation of "messageQueue" and "uiMessageQueue" made it unclear that both point towards the UI thread, and "messageQueue" wasn't the normal NM queue.