Better Bundled Test Disovery
Integration tests are run on the RN instance instance of the Test runner. We want the test runner to be able to discover tests that were defined. We currently do that statically, by looking at the AST of tests and parsing out what test components are registered.

This change makes a few changes to how these are represented:
1. We look for CallExpressions beyong file-scope, removing some restrictions
2. We group tests into describe blocks on the runner side based on the source file they appear in
3. Since this makes naming collisions less obvious, we wnsure on the RN side we don't register multiple tests with the same component name (this will lead to loss of coverage)

An alternative architecture that might be worth pursuing in the future is to let the RN instance know about test structure, and to query it from the test runner. This is more flexible, and would allow smarter assignment of components, but is also a bit more work to implement.