Ensure Integration Tests Show Exceptions When Source Resolution Fails (#6047)
RN will sometimes fail to prettify stack information with source map information. In this case, the RedboxHandler never sees a call to update error information. We don't have a great signal of when this fails, so this change alters logic so that a timeout will show exception information if we have an unresolved exception pending.

Validated cases around:
- Tests passing
- Badly behaved tests calling testPassed multiple times not leaking
- Live reload
- Exceptions where resolution succeeds
- Exceptions where resolution fails