Supply UniversalApiContract for Platform.Version (#6074)
* Supply UniversalApiContract for Platform.Version

Fixes #4176

When implementing PlatformConstants I originally took the route of not providing a version, since Windows likes to hide it. Version is officially part of typings though, and it seems some code will be unhappy if its undefined. With this change, we provide UniversalApiContract as a proxy of "what native functionality is available", similar to how Android Platform versions work.

This module is shared between desktop/XAML, which has some spcial considerations. Namely when running on Windows < 10, we should just return 0 for API contract.

Added integration tests to validate this is working correctly.

* Change files