Free Up Some Memory for react-native-init CI Jobs (#6111)
* Free Up Some Memory for react-native-init CI Jobs

For the integration test job I discovered we get  more reliable builds if we do build-work when daemons are not active. For react-native-init tests, we don't need Verdaccio by the time we're ready to build, so we can kill it to reclaim a bit of memory.

Testing locally, Verdaccio without any packages seems to use ~50MB while idle, which is enough to be worth reclaiming, but less than I had hoped for. Not sure if the number would be different when packages are present.

* Reconcile with later task