Integrate 7/29 Nightly Build
Commits: https://github.com/facebook/react-native/compare/e96f1e1d8...949dedca9

(not adding a changelog for the description on this one since there are 120 commits)

Doing a larger integration than usual because web debugging was broken upstream from 7/9 to 7/28.

Not many changes that concern us apart from some JS dependency changes. Some of the changes pulled out into monorepo packages may effect us if changes happen to them that are unpublished, and we will need a better story for those soon.

We need to make local copies of flow typings that are currently in RN core but not published. Starting up a discussion there and tracking with #6139