Fix WinUI3 continueOnBuildError
The last change added the flag to the matrix, but we weren't passing from the matrix to a template parameter.

Annoyingly everything in the matrix is exposed as a runtime string, so we cannot simply convert to boolean before passing to the init template, since runtime expressions are limited to conditons and variables. Switching to loosely typed parameters, I saw all sorts of weird beahavior going from Matrix to parameter. Separating out the WimUI3 builds into a separate job and keeping strongly typed parameters seems to be a less fragile path.

Validated we get expected results by artitifically creating build failures in the pipeline.