Integrate 8/5 Nightly RN Build (#6145)
* Integrate 8/5 Nightly RN Build

Commits: https://github.com/facebook/react-native/compare/949dedca9...10b4b9505

Interesting bits for us are some foundational TurboModule work to pass along codegen scheme when obtaining a TurboModule from TurboModuleManager. This is a noop in core right now and it for us to.

The other bit that we needed to follow along with was some renaming for babel-plugin-inline-view-configs to @react-native/babel-plugin-codegen.

* Change files