Integrate 8/17 Nightly RN Build
Commits: https://github.com/facebook/react-native/compare/10b4b9505...423155127

Going ahead to this build, since it moves us off of an experimental version of React that confuses component governance. The interesting bits are new APIs for NativeAnimatedModule. No regression if we don't implement them, and they're not implemented on iOS, so this is logged as #6161 for future work.