Allow Default baseVersion in Override Manfests
Fixes #5777

We see a bunch of noisiness in each integration PR due to the same >100 lines changing in override manifests. This change allows manifests to declare a default react-nartive version to reduce the number of diffs we see.

Manually tested overide validation. Still need to test the upgrade path and add unit + e2etests for the changes.