Remove the other TimingModule Assertion
TimingModule methods are dispatched to the UI thread. The instance/Module may be gone by the time its methods are processed in the UI queue. We removed an assertion that an instance is available on render callback, but there is a second assertion on method call as well. Remove the assertion.