Allow Default baseVersion in Override Manifests (#6165)
* Allow Default baseVersion in Override Manfests

Fixes #5777

We see a bunch of noisiness in each integration PR due to the same >100 lines changing in override manifests. This change allows manifests to declare a default react-nartive version to reduce the number of diffs we see.

Manually tested overide validation. Still need to test the upgrade path and add unit + e2etests for the changes.

* Change files

* Update upgradeOverrides e2etest with new behavior

* Use new scheme during manifest generation

* Add UTs

* Fix merged baseVersion