Force hermes-engine 0.70
The version of Hermes used by RN 0.63 up to recent RN nightlies is vulerable to CVE-2020-1912, which shows up in component governance alerts. We don't actually use the npm package binary in any way, so this change pins a newer version of Hermes until we integrate the 9/17 Nightly RN build that will be part of RN 0.64.