Use newer version of appium-support
We currently bring in an old version of appium-support which depends on the md5-file dependency that triggers component governance. Newer versions of appium-support should theoretically remain compatible with its dependants semver-wise. This change makes it so we bring in the newer version.

It basically entailed manually clearing out chunks of the lockfile and telling yarn to come up with new resolutions, since yarn seems to not support directly upgrading indirect dependencies correctly in our repo.