Drop Issue Requirement from Override Manifest Schema Validation (#6263)
* Drop Issue Requirement from Override Manifest Schema Validation

Most types of overrrides require an issue number. This is done on the schema level, and is done incosistently per-type, where some allow no issues or LEGACY_FIXME, where others allow none. This was done to add barriers to forking, but ends up being a bit to broad/heavy-handed, and the differences between requirements of override types is confusing.

This changes makes the schema uniform, dropping the special LEGACY_FIXME issue and universally allowing missing issues.

To guide engineers to the right path, we still do require issues depending on types in the override prompt, but these can be altered manually.

* Change files

* Update copies to undefined issue during manifest generation

* serializeBase cleanup