Fix Crash when Logbox shown on pre-19H1
ActualHeight was added to UIElement in 19H1, but is being used in native code for Logbox/legacy redbox. This will also cause integration tests to crash on redbox, since they're running on debug build on Windows Server 2019.

Replace usages with ActualHeight/ActualWidth.

Initial thoughts are to not backport this, since it's dev only and has existed for a couple of versions.