Allow E2ETest to Work Without Back Button (And lots of cleanup) (#6344)
* Allow E2ETest to Work Without Back Button (And lots of cleanup)

An upcoming RNTester change removes the back button. E2ETest relies on this as a signal that a test page has been loaded, so we need to replace it with something else. In this case we use the absence of the RNTesterList search box. Note that a search box may stilll be present, but will have a different test id than the RNTesterList one.

Now that we have visual separation between component examples and API examples, we update HomePage to let it know the difference instead of keeping that logic just in the VisitAllPages test. Doing this led to some discoveries that led to quite a bit of house-cleaning.

Verified everything runs locally, and that artifiically created tree dump failures show correctly.

* PR feedback

* Disable macOS tests