Allow Hermes to be used in C++ projects without modifying source code in NPM package (#6350)
* Allow Hermes to be used in C++ projects without modifying source code

[Using hermes](https://microsoft.github.io/react-native-windows/docs/hermes) on Windows currently requires altering source code, preventing its usage for consumption scenarios. This change enables its usage without modifying source code or the public NPM package.

This is accomplished by parameterizing the build of MSRN based on a new "BuildFlags" property sheet in the user solution. This mechanism replaces the one used for WinUI3, of dropping an empty "UseWinUI3" file that the build looks for. We also need to enable restoration of the Hermes NuGet package before the MSRN build. We can do this for C++ projects by adding to the user packages.config which is restored to the solution before building. We cannot use the same trick for C# apps, which use the PackageReference style of restoration during build. This would require us to add Hermes uncondtionally to the MSRN packages.config, which would add a large binary cost to all, or to switch the C# project to pre-build restoration. For now, we just punt on C# projects, which are less likely to be interested in bleeding edge perf anyway.

New projectrs can be created to use Hermes now by adding the "--useHermes" flag during project init. I plan to add this to docs after merging.

Note that parameterizing the build won't work long term, and isn't the approach we will want to take with binary distribution. Ideally we will have an ABI safe JSI runtime interface so we can plugin in the runtime without needing to build different versions of MSRN, but we don't have what we need to accomplish this quite yet.

Locally validated we can launch a newly created C++ app and it shows the Hermes badge in the new app screen with web debugging disabled. Will validate a WinUI3 build before merging.

* Change files

* whitespace

* Import BuildFlags in a few more places

* Set default JSIEngineOverride baed on preprocessor definition

* Fix typo

* Don't add Hermes targets to template project (MSRN already imports the targets)

* typo