Extend Playground Metro Config to Allow Live Reload of RNTester Changes (#6365)
ACoates added a custom metro config for Playground which attempts to resolve source files for react-native-windows, which are normally moved to a new target location during the build. RNTester sources were moved to their own package, and were not being included by the custom metro config. This change updates the metro config to allow it to look in arbitray "dev packages" where it will prefere JS sources in src instead of the root. @react-native-windows/tester is included as a dev package to let its sources be resolved.

Validated modifying src files in vnext and @react-native-windows/tester and observed we see fast refresh.