Fix Generator Failing to Overwrite Template Files
Fixes #6380

During project initialization, we prompt to ask users whether to replace certain files, like the default metro config. We used to have a custom prompt to do this, but replaced it with Inquirer in 0.63. A typo + lack of type safety from inquirer means that we  regardless of user response, we treat `shouldReplace` as undefined which is coerced to false.

Validated the stock metro config is correctly replaced after the change.