[0.63] Fix crash during instance destruction if there are outstanding animations (#6321)
We were never notifying JS to detach a root view when the rootview went away. So JS would not send commands to clean up animations from that rootview.

There was an issue in ReactRootControl where it would clear out the UI the JS created as part of reload, which would cause a crash when the JS then tried to remove that UI.

We were not properly quitting the main UI queue, so things would get run there after the instance was gone.