Integrate 10/5 Nightly RN Build
Commits: https://github.com/facebook/react-native/compare/9b094ee77...77f7658e4

The bits that effect us are mostly mechanical changes. The Text component had a decent chunk of merge conflicts with our forking of it to add border props to the JS. Validated Text borders, etc in Text RNTester page.

https://github.com/facebook/react-native/commit/7c93f5b0015306b08f59f07254ea99c81a673135 moved the location of TurboModule C++ code, so we need to update out native build logic to accomadate it.

The TurboModuleExample is now conditionally loaded based on the presence of __turboModuleProxy JS side. Requiring this exposes to Metro that we run into some of the same issues with relative imports escaping the RNTester package as seen before, that is fixed in a later RN build. We patch around the import while waiting for a new update.

Manually validated against RNTester in XAML + NetUI (areas touched + new examples).