Fix Bundling With "react-native-windows-init --useDevMode"
Fixes #5605

--useDevMode operates using `yarn link`, which symlinks the `react-native-windows` directory in node_modules. This confuses Metro, which doesn't understand symlinks, and doesn't know how to find modules hoisted above `react-native-windows`.

Copy a custom metro config when using `--useDevMode` so we can bundle in the symlinked environmment.

Validated creating a new app, validating we have the devmode config, then launching the app.

Note that we still run into cross-package asset issues, where we're missing an image from the new app screen. This isn't super needed for the new app screen, but maybe something we should fix as well.