Update Microsoft.ReactNative Nuget Package to support layout convention (#6381)
NuGet defines a layout semantic for UWP projects. CSharp projects use the directory
convention which is to copy all dll's and pdb recursivley under the 'native' folder
This means that if we have native\release\foo.dll and native\debug\foo.dll
UWP will try to copy both versions of foo.dll to the output app and will
collide. Cpp projects don't have this problem because they don't support Nuget
properly, and require the producing nuget package to ship a targets file
that overrides the behavior manually where we accounted for.

This change drops including the debug bits in Microsoft.ReactNative.
Just like the Microsoft.ReactNative.Managed package if a debug slice is build for a
platform without a release versions, that debug version will be included.