Avoid double transpilation of TS overrides
Metro with its default preset can already transform TS through babel, so there is no need to run the TypeScript compiler as well during the build. This change uses builtin Metro logic for transpilation, which also ensures the devMode metro config produces indentical results to production.

We remove tsc compilation from the build to save time, but add it as a linting step with noEmit to check for type errors.