Remove Build-Time Native Code Patching
Fixes #5168

When moving of of the microsoft/react-native fork, quite a bit of native code had differences. To stage these changes, we adding patching of native code at build time to replace it with our own copies. This patching was implemented by constructing a file tree with only the patched copies, and compiling from that as the source of truth. This was done instead of include directory manipulation to ensure that locally included headers were correctly patched.

Patches have been reduced to a single change to Yoga, which we've been unable to upstream or work around. Because we don't have headers, we can use a simpler approach that avoids the patching. We haven't needed additional native patches over the last year, so it's likely these limitations should hold up.

This change removes the patching step, in favor of treating the patched Yoga.cpp as a one-off.